running pretraining

+++ preprocessing/run_exp.py
@@ -8,17 +8,19 @@ def get_all_files(path):
 
 
 if __name__ == '__main__':
-    for lang in langs:
-        path = '/mnt/c/TransCoder/data/WASM/' + lang + '/'
-        for file_name in get_all_files(path):
-            if not file_name.endswith('corpus_small_fixed'):
-                os.system('rm ' + path + file_name)
-    print('killed files')
-    os.system('rm -rf /mnt/c/TransCoder/data/WASM/c-wat-')
-    print('killed combined')
+    long = False
+    if long:
+        for lang in langs:
+            path = '/mnt/c/TransCoder/data/WASM/' + lang + '/'
+            for file_name in get_all_files(path):
+                if not file_name.endswith('corpus_small_fixed'):
+                    os.system('rm ' + path + file_name)
+        print('killed files')
+        os.system('rm -rf /mnt/c/TransCoder/data/WASM/c-wat-')
+        print('killed combined')
 
-    print('start to_json')
-    os.system('python3 /mnt/c/TransCoder/preprocessing/to_json.py')
+        print('start to_json')
+        os.system('python3 /mnt/c/TransCoder/preprocessing/to_json.py')
 
     print('start training')
     os.system('python3 -m preprocessing.preprocess  /mnt/c/TransCoder/data/WASM --lang1 c --lang2 wat --bpe_train_size 0 --test_size 100 --local True')
